requirements.txt
pandas
openpyxl
rapidfuzz
Flask>=3.0
Flask-SQLAlchemy>=3.1
SQLAlchemy>=2.0
Flask-Mail>=0.9.1
Flask-WTF>=1.2.1
python-dotenv>=1.0
gunicorn>=21.2.0
Flask-Limiter>=3.5.0
Flask-Talisman>=1.1.0
Flask-Compress>=1.14
boto3>=1.34.0
psycopg2-binary>=2.9.9
Werkzeug>=3.0
Pillow>=10.0.0
requests>=2.31.0
